requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
Pillow==9.5.0
pyzbar==0.1.9
requests==2.30.0
urllib3==2.0.2
